+++ build-release.ps1
@@ -1,5 +1,5 @@
 param(
-    [string]$Version = "2.2.16",
+    [string]$Version = "3.0.0",
     [string]$OutputDirectory
 )